Move `Unfollow` command
